Refactor: Make registerGhostCells a member function of AIManager.
This way we could potentially store state in AIManager if needed.
